further cobra-ifying the CLI
