+++ cmd/cli/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"errors"
 	"io"
 	"log"
 	"os"
@@ -16,33 +15,11 @@ const (
 
 func newRootCmd(out io.Writer, conn *grpc.ClientConn) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "goprox",
-		Short: "The Goprox dependency manager",
-		// Long:         globalUsage,
+		Use:          "goprox",
+		Short:        "The Goprox dependency manager",
 		SilenceUsage: true,
-		// PersistentPostRun: func(cmd *cobra.Command, args []string) {
-		// 	teardown()
-		// },
 	}
-	// p := cmd.PersistentFlags()
 	cmd.AddCommand(newGetCommand(out, conn))
-	return cmd
-}
-
-func newGetCommand(out io.Writer, conn *grpc.ClientConn) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "get PACKAGE VERSION",
-		Short: "download a package to your vendor directory",
-		// Long:  createDesc,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) != 2 {
-				return errors.New("package & version are required")
-			}
-			name, version := args[0], args[1]
-			return get(out, conn, name, version)
-		},
-	}
-
 	return cmd
 }
